+++ src/Contracts/HasPasskeysAuthentication.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace Happenv\FilamentMultiFactorPasskeys\Contracts;
+
+use Laravel\Passkeys\Contracts\PasskeyUser;
+
+interface HasPasskeysAuthentication extends PasskeyUser {}